Revert "Simplify -saveDataWithBlock: to use -MR_contextForCurrentThread"
This was a *really* bad idea. Using NSThread's -threadDictionary in
conjunction with managed object contexts and dispatch queues brought out
everything possibly wrong with threading, painful to debug deadlocks and
race conditions in particular.

The cause was that the threadDictionary's contents, including the
per-thread context, were being cleared by the dispatch queue between
block dispatches. The timing of this is what caused all the problems,
there was no way to control when the context, and thus the whole core
data stack, were being called into, nor what state they would be in.

The safe and reliable way is to do this how it was previously done,
create a new context for each background block. Comments have been added
to this commit to reflect this.

This reverts commit 82c737c9aa0d5e63c09a655efee3a5feff65b491.