Skip completion block in absence of default context
It's not documented, but the default context should exist at the time
the completion callback. If it doesn't, the callback is never run. This
scenario can potentially happen when the application is resumed. Do not
write code in the completion callback the needs to guaranteed to run
during a suspend / resume cycle.